project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/configchange/ComponentConfigChangeRTMConsumable.java
package com.smartsparrow.rtm.subscription.courseware.configchange;

import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.ConfigChangeBroadcastMessage;
import com.smartsparrow.rtm.subscription.data.AbstractRTMConsumable;
import com.smartsparrow.pubsub.data.RTMEvent;

/**
 * This RTM consumable describes an component config change event
 */
public class ComponentConfigChangeRTMConsumable extends AbstractRTMConsumable<ConfigChangeBroadcastMessage> {

    private static final long serialVersionUID = 2087229202974132955L;

    public ComponentConfigChangeRTMConsumable(RTMClientContext rtmClientContext, ConfigChangeBroadcastMessage content) {
        super(rtmClientContext, content);
    }

    @Override
    public RTMClientContext getRTMClientContext() {
        return rtmClientContext;
    }

    @Override
    public String getName() {
        return String.format("author.activity/%s/%s", content.getActivityId(), getRTMEvent().getName());
    }

    @Override
    public String getSubscriptionName() {
        return ActivityRTMSubscription.NAME(content.getActivityId());
    }

    @Override
    public RTMEvent getRTMEvent() {
        return new ComponentConfigChangeRTMEvent();
    }

    @Override
    public boolean equals(Object o) {
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return super.hashCode();
    }
}
